Lots of changes related to signing and exceptions.
All checked exceptions are now of type PassSigningException, with the
underlying exception (if any) included as the cause.

PassSignerImpl and its Builder:
PassSignerImpl now constructs its CMSSignedDataGenerator upon its
creation, instead of with each invocation of generateSignature.
KeyStore aliases for the private key and signing certificate can be set
separately.
Deprecated the keystore methods which included alias as a parameter in
favor of setting it with the new method alias.

PassSigningUtil:
firstAlias method throws exceptions if there are not exactly 1 aliases
in the KeyStore.