Fixed bug caused by Include.NON_DEFAULT in new version of Jackson
NON_DEFAULT doesn't seem to be working in the newer version.
In the case of Location.altitude (which is optional), not setting a
value results in the field not being ignored as it appearing in the
pass.json file:

{
    "latitude" : 37.33182,
    "longitude" : -122.03118,
    "altitude" : "NaN"
}

This is not a valid format which results in the Wallet app not reading
the pass.

I changed Include.NON_DEFAULT to Include.NON_NULL and set the default
altitude value to null. This results in the same behaviour as before
the upgrade of Jackson.